backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openai==1.3.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
PyPDF2==3.0.1
requests==2.31.0 